requirements.txt
# ── Core ──────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
pydantic==2.7.1

# ── Google Cloud ──────────────────────────────────────────────
google-cloud-bigquery==3.23.1
google-cloud-aiplatform==1.57.0
google-generativeai==0.7.0

# ── RAG / Embeddings ──────────────────────────────────────────
langchain==0.2.3
langchain-google-genai==1.0.6
langchain-community==0.2.4
pypdf==4.2.0
numpy==1.26.4
faiss-cpu==1.8.0          # Local vector store (swap for AlloyDB later)

# ── Data ──────────────────────────────────────────────────────
pandas==2.2.2
pyarrow==16.1.0

# ── Dev ───────────────────────────────────────────────────────
httpx==0.27.0             # For testing FastAPI endpoints
